+++ config/app.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 return [
     'name' => 'Jasanika 2',
-    'version' => '0.21',
+    'version' => '0.22',
     'debug' => false,
     'env' => 'production',
 ];